% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APAreport.R
\name{APAreport}
\alias{APAreport}
\title{APAreport}
\usage{
APAreport(toc = TRUE, year, publisher = "PUBLISHER", preamble, fancyHeader,
  latex_engine = "xelatex")
}
\arguments{
\item{toc}{Logic variable indicating if Table of Contents should be included. Type it in YAML header.}

\item{year}{Is the year of the publication. Its calculated from the function \code{Sys.time}. Type it in YAML header.}

\item{publisher}{The publisher of the document.Type it in YAML header.}

\item{preamble}{Latex preamble if you wish to include further features to the template. Type it in YAML header.}

\item{fancyHeader}{Logic Variable indicating if fancy header, from \code{fancyhdr} Latex package, should be used. Default TRUE. Type it in YAML header.}

\item{latex_engine}{Inherit from \code{bookdown::pdf_document2}. Change the default to Xelatex}

\item{...}{Arguments to \code{bookdown::pdf_document2}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Format for creating a technical report. It is asummed that the publisher is the institution or institutions that produced the report, as no affiliation field is allowed for authors. Adapted
from
\href{http://www.LaTeXTemplates.com}{http://www.LaTeXTemplates.com}.
By  Peter Wilson (herries.press@earthlink.net). Under licence CC BY-NC-SA 3.0 \href{http://creativecommons.org/licenses/by-nc-sa/3.0/}{http://creativecommons.org/licenses/by-nc-sa/3.0/},
It also inherits parameters from \code{pdf_document2} from \pkg{bookdown}, by Yihui Xie et al., released under license GPL-3. Many of the preamble is based on the \code{latex.default} from \pkg{rmarkdown}. You may want to check the basics of templates \href{here}{http://ismayc.github.io/ecots2k16/template_pkg/} the \href{default.latex}{https://github.com/rstudio/rticles/blob/master/inst/rmarkdown/templates/ctex/resources/default.latex}.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "APAreport", package = "DMtemplates")
}
}
